test(desktop): shared fixtures
Temporary directories, microtask drains, fake clocks, native-helper child
processes, an in-memory brain envelope and an AppSettingsView literal were
each hand-rolled in several test files, and every copy had drifted: three
temporary directories were never cleaned up and four drains had settled on
four different tick counts for the same wait.

One copy of each lands under the existing `#testing/*` alias, beside the
fixtures already there, with an AGENTS.md naming what each is for and a
repository check that fails the build on a hand-rolled `mkdtemp` or
`setImmediate` under `apps/desktop/src`.

No test name changes and no assertion moves: every signature edit is
`async () =>` becoming `async (t) =>` so a helper can take a TestContext.
