+++ apps/desktop/src/main/brain/host-lifecycle.test.ts
@@ -3,280 +3,33 @@ import test from "node:test";
 import {
   BRAIN_REQUEST_ORIGIN,
   BRAIN_REQUEST_STATUS,
-  BrainAgent,
-  type BrainAgentOptions,
   type BrainRequestRecord,
-  type BrainStateStorage,
-  BrainStateStore,
   brainStateFromStored,
-  responsesModelAnswer,
 } from "@sidecar/brain";
-import { brainReplyWords, isTerminalBrainRequestStatus } from "@sidecar/brain/requests";
+import { brainReplyWords } from "@sidecar/brain/requests";
+import { CONVERSATION_ENTRY_KIND } from "@sidecar/realtime";
+import { QUEUE_MODE } from "@sidecar/runtime";
 import {
-  type BareResponsesModel,
-  bareModelAdapter,
-  toolLoopRuntimeOver,
-} from "@sidecar/brain/testing";
-import {
-  appendConversationThreadEntry,
-  CONVERSATION_ENTRY_KIND,
-  type ConversationEntry,
-} from "@sidecar/realtime";
-import { DeliveryLedger, type DeliveryRecord, QUEUE_MODE } from "@sidecar/runtime";
-import { DELIVERY_STATE, type ModelResponse } from "@sidecar/runtime-contracts";
-import { ACT_RESULT_STATUS, type WireRecord } from "@sidecar/wire";
-import type {
-  BrainReplyClaimResult,
-  BrainReplyOffer,
-  BrainRequestSnapshot,
-} from "#shared/messages/brain";
-import { deliverable, type GrantedWords, ledgerContext } from "../gateway/service";
-import { operatorOverBrain } from "../gateway/testing";
-import { VoiceReceiver } from "../voice-receiver";
-import { BrainHost } from "./host";
-import { followBrainRequests } from "./ipc";
-
-/**
- * The real agent, store, host, follower, and submission path composed as the
- * main process composes them, with only the model and the disk synthetic.
- */
-
-const NOW = 1_800_000_000_000;
-
-class MemoryStorage implements BrainStateStorage {
-  file: string | undefined;
-  read() {
-    return this.file;
-  }
-  write(contents: string) {
-    this.file = contents;
-    return true;
-  }
-}
-
-/** A model that answers nothing until the test says so. */
-function heldClient(): BareResponsesModel & { release: (answer: ModelResponse) => void } {
-  const waiting: ((answer: ModelResponse) => void)[] = [];
-  return {
-    respond: () =>
-      new Promise((resolve) => {
-        waiting.push(resolve);
-      }),
-    quietUntil: () => undefined,
-    release: (answer) => {
-      for (const resolve of waiting.splice(0)) resolve(answer);
-    },
-  };
-}
-
-function settle(): Promise<void> {
-  return new Promise((resolve) => {
-    let ticks = 0;
-    const tick = () => {
-      ticks += 1;
-      if (ticks > 30) resolve();
-      else setImmediate(tick);
-    };
-    tick();
-  });
-}
-
-function composed() {
-  const storage = new MemoryStorage();
-  let ids = 0;
-  const store = new BrainStateStore({
-    storage,
-    createGenerationId: () => `gen-${++ids}`,
-    now: () => NOW,
-  });
-  let thread: readonly ConversationEntry[] = [];
-  let refuseWrites = false;
-  const broadcasts: (readonly BrainRequestSnapshot[])[] = [];
-  const record = (entry: ConversationEntry, at: number) => {
-    if (refuseWrites) return false;
-    thread = appendConversationThreadEntry(thread, entry, NOW + 1000, at);
-    return true;
-  };
-  // The delivery owner and the receiver, composed as desktop-app composes
-  // them: every report is observed before it is broadcast, every published
-  // end is offered to a ready receiver, and readiness flushes what waited.
-  const deliveries = new DeliveryLedger<GrantedWords>({
-    nextDeliveryId: () => `delivery-${++ids}`,
-  });
-  const receiver = new VoiceReceiver();
-  const offers: BrainReplyOffer[] = [];
-  const offerReplies = () => {
-    if (!receiver.isReady()) return;
-    const offer = deliveries.nextOffer(receiver.epoch());
-    if (offer) offers.push(offer);
-  };
-  receiver.onReady(offerReplies);
-  store.onReplaced(() => deliveries.reset());
-  const host = new BrainHost({
-    follow: (agent) =>
-      followBrainRequests(agent, {
-        recordConversationEntry: record,
-        broadcastRequests: (snapshots) => {
-          deliveries.observe(
-            snapshots.map((snapshot) => ({
-              runId: snapshot.runId,
-              ended: isTerminalBrainRequestStatus(snapshot.status),
-            })),
-          );
-          broadcasts.push(snapshots);
-        },
-        onEndPublished: (ended) => {
-          const generationId = store.generationId();
-          if (generationId !== undefined && deliverable(ended)) {
-            deliveries.published(ended.runId, generationId);
-          }
-          offerReplies();
-        },
-      }),
-    publishEmpty: () => broadcasts.push([]),
-  });
-  /** Submits as a window does, through the operator over the standing brain; the ask's own line is written here. */
-  const { submit } = operatorOverBrain({
-    current: () => host.current(),
-    recordConversationEntry: record,
-  });
-  const claimContext = () => ({
-    receiverCurrent: (epoch: number) => receiver.isReady() && receiver.epoch() === epoch,
-    generationStands: (generationId: string) => store.holdsGeneration(generationId),
-    liveRecord: (runId: string) => host.current()?.request(runId),
-  });
-  const claim = (offer: BrainReplyOffer): BrainReplyClaimResult => {
-    const granted = deliveries.claim(
-      offer.runId,
-      offer.deliveryId,
-      offer.epoch,
-      ledgerContext(claimContext()),
-    );
-    return granted.granted
-      ? { granted: true, words: granted.words.words, origin: granted.words.origin }
-      : { granted: false };
-  };
-  /** Queued or offered: everything owed that no receiver has taken in hand. */
-  const unclaimed = (): readonly DeliveryRecord[] =>
-    deliveries
-      .records()
-      .filter(
-        (delivery) =>
-          delivery.state === DELIVERY_STATE.QUEUED || delivery.state === DELIVERY_STATE.OFFERED,
-      );
-  const acknowledge = (offer: BrainReplyOffer) => {
-    const emptied = deliveries.acknowledge(offer.runId, offer.deliveryId, offer.epoch);
-    if (emptied) offerReplies();
-    return emptied;
-  };
-  /** The wait as main's IPC answers it: publication let finish, live record re-read, the call granted or not. */
-  const waitOnCall = async (runId: string, epoch: number, timeoutMs = 1) => {
-    const agent = host.current();
-    const waited = await agent?.waitAsk(runId, timeoutMs);
-    if (!waited || !isTerminalBrainRequestStatus(waited.status))
-      return { record: waited, speak: false };
-    await settle();
-    const live = agent?.request(runId) ?? waited;
-    if (live.historyRecordedAt === undefined) return { record: live, speak: false };
-    const generationId = store.generationId() ?? "";
-    const speak =
-      deliverable(live) &&
-      deliveries.grantOnCall(live.runId, generationId, epoch, ledgerContext(claimContext()));
-    if (speak) offerReplies();
-    return { record: live, speak };
-  };
-  const build = (client: BareResponsesModel, options: Partial<BrainAgentOptions> = {}) => {
-    const model = bareModelAdapter(client);
-    return new BrainAgent({
-      ...options,
-      runtime: toolLoopRuntimeOver(model),
-      prepareTurn: () => ({ prompt: "instructions", layers: {} }),
-      acts: { perform: async () => ({ status: ACT_RESULT_STATUS.ACCEPTED }) },
-      roster: () => ({ text: "", identities: [] }),
-      standingContext: () => "",
-      readTranscriptSince: async () => ({ status: ACT_RESULT_STATUS.REJECTED, reason: "no" }),
-      readTranscript: async () => ({ status: ACT_RESULT_STATUS.REJECTED, reason: "no" }),
-      deliver: () => undefined,
-      store,
-      createRunId: () => `run-${++ids}`,
-      report: () => {},
-      now: () => NOW,
-      // A run left in flight at a test's end must not hold the process open:
-      // its deadline timers are unreferenced, as the test's own would be.
-      schedule: (callback, delayMs) => {
-        const timer = setTimeout(callback, delayMs);
-        timer.unref();
-        return timer;
-      },
-      // SAFETY: the handle is what `schedule` above returned, which is always a `setTimeout` timer.
-      cancel: (timer) => clearTimeout(timer as ReturnType<typeof setTimeout>),
-    });
-  };
-  return {
-    storage,
-    store,
-    host,
-    build,
-    record,
-    submit,
-    thread: () => thread,
-    broadcasts,
-    deliveries,
-    unclaimed,
-    receiver,
-    offers,
-    claim,
-    acknowledge,
-    waitOnCall,
-    refuse: (value: boolean) => {
-      refuseWrites = value;
-    },
-  };
-}
-
-/** A raw Responses payload as the adapter would normalize it. */
-function answerOf(payload: WireRecord): ModelResponse {
-  const answer = responsesModelAnswer(payload);
-  assert.ok(answer);
-  return answer;
-}
-
-function answered(text: string): ModelResponse {
-  return answerOf({
-    output: [{ type: "message", role: "assistant", content: [{ type: "output_text", text }] }],
-  });
-}
-
-function replies(thread: readonly ConversationEntry[], runId: string) {
-  return thread.filter((e) => e.kind === CONVERSATION_ENTRY_KIND.REPLY && e.requestId === runId);
-}
-
-async function submitMany(submit: ReturnType<typeof composed>["submit"], count: number) {
-  const runIds: string[] = [];
-  for (let index = 0; index < count; index += 1) {
-    const result = await submit({
-      submissionId: `sub-${index}`,
-      question: `ask ${index}`,
-      origin: BRAIN_REQUEST_ORIGIN.TYPED,
-    });
-    assert.equal(result.outcome, "accepted");
-    if (result.outcome === "accepted") runIds.push(result.runId);
-  }
-  return runIds;
-}
+  answered,
+  answerOf,
+  brainHarness,
+  heldModel,
+  BRAIN_HARNESS_NOW as NOW,
+} from "#testing/brain-harness";
+import { drainMicrotasks } from "#testing/drain";
 
 test("removing the capability under five outstanding runs leaves every run interrupted, published, and marked", async () => {
-  const c = composed();
-  const client = heldClient();
+  const c = brainHarness();
+  const client = heldModel();
   await c.host.replace(() => c.build(client));
   const agent = c.host.current();
   assert.ok(agent);
-  const runIds = await submitMany(c.submit, 5);
-  await settle();
+  const runIds = await c.submitMany(5);
+  await drainMicrotasks();
   assert.equal(c.thread().filter((e) => e.kind === CONVERSATION_ENTRY_KIND.TYPED_ASK).length, 5);
 
   await c.host.replace(() => undefined);
-  await settle();
+  await drainMicrotasks();
   assert.equal(c.host.current(), undefined);
   const stored = brainStateFromStored(c.storage.file);
   assert.equal(stored?.requests.length, 5);
@@ -303,7 +56,7 @@ test("removing the capability under five outstanding runs leaves every run inter
       ],
     }),
   );
-  await settle();
+  await drainMicrotasks();
   assert.equal(c.thread().filter((e) => e.words === "late").length, 0);
   assert.equal(
     brainStateFromStored(c.storage.file)?.requests.every(
@@ -314,19 +67,19 @@ test("removing the capability under five outstanding runs leaves every run inter
 });
 
 test("a successor replacing the agent under outstanding runs inherits a thread with every end written, and owns the store alone", async () => {
-  const c = composed();
-  const first = heldClient();
+  const c = brainHarness();
+  const first = heldModel();
   await c.host.replace(() => c.build(first));
   const agent = c.host.current();
   assert.ok(agent);
-  const runIds = await submitMany(c.submit, 5);
+  const runIds = await c.submitMany(5);
   // An earlier publication is held: the thread refuses until the handoff.
   c.refuse(true);
-  await settle();
+  await drainMicrotasks();
   c.refuse(false);
-  const second = heldClient();
+  const second = heldModel();
   await c.host.replace(() => c.build(second));
-  await settle();
+  await drainMicrotasks();
   const successor = c.host.current();
   assert.ok(successor && successor !== agent);
   for (const runId of runIds) {
@@ -346,7 +99,7 @@ test("a successor replacing the agent under outstanding runs inherits a thread w
   assert.equal(result.outcome, "accepted");
   // The host reaches its first inference only after the run's own
   // bookkeeping; the answer is released once the model has been asked.
-  await settle();
+  await drainMicrotasks();
   second.release(
     answerOf({
       output: [
@@ -354,7 +107,7 @@ test("a successor replacing the agent under outstanding runs inherits a thread w
       ],
     }),
   );
-  await settle();
+  await drainMicrotasks();
   const fresh = brainStateFromStored(c.storage.file)?.requests.find(
     (r) => r.question === "new ask",
   );
@@ -377,15 +130,15 @@ test("a successor replacing the agent under outstanding runs inherits a thread w
 });
 
 test("a reset under outstanding runs discards them without publishing, and the successor starts clean", async () => {
-  const c = composed();
-  const client = heldClient();
+  const c = brainHarness();
+  const client = heldModel();
   await c.host.replace(() => c.build(client));
   const agent = c.host.current();
   assert.ok(agent);
-  await submitMany(c.submit, 3);
-  await settle();
+  await c.submitMany(3);
+  await drainMicrotasks();
   assert.equal(await c.store.clear(), true);
-  await settle();
+  await drainMicrotasks();
   assert.deepEqual(agent.requests(), []);
   assert.equal(c.thread().filter((e) => e.kind === CONVERSATION_ENTRY_KIND.REPLY).length, 0);
   // The file holds the empty successor and the marker of the erasure alone.
@@ -393,19 +146,19 @@ test("a reset under outstanding runs discards them without publishing, and the s
   assert.equal(stored?.requests.length, 0);
   assert.equal(stored?.reset?.generationId, "gen-1");
   await c.host.replace(() => undefined);
-  await settle();
+  await drainMicrotasks();
   assert.equal(c.thread().filter((e) => e.kind === CONVERSATION_ENTRY_KIND.REPLY).length, 0);
 });
 
 test("a run ending long after its wait is offered once, to a ready receiver, only after its line and mark landed", async () => {
-  const c = composed();
-  const client = heldClient();
+  const c = brainHarness();
+  const client = heldModel();
   await c.host.replace(() => c.build(client));
   const agent = c.host.current();
   assert.ok(agent);
   const epoch = c.receiver.begin();
   c.receiver.markReady(epoch);
-  const [runId] = await submitMany(c.submit, 1);
+  const [runId] = await c.submitMany(1);
   assert.ok(runId);
   // The wait comes back pending — the ask outlived its thirty seconds — and
   // the developer's call is released with no grant. Nothing is owed yet.
@@ -416,8 +169,8 @@ test("a run ending long after its wait is offered once, to a ready receiver, onl
   // marked, and the reply waits for a receiver rather than landing on none.
   c.receiver.reset();
   client.release(answered("Two agents are waiting."));
-  await settle();
-  assert.equal(replies(c.thread(), runId).length, 1);
+  await drainMicrotasks();
+  assert.equal(c.replies(runId).length, 1);
   assert.equal(c.unclaimed().length, 1);
   assert.equal(c.offers.length, 0);
   // Readiness flushes it, exactly once, under the epoch that reported.
@@ -432,18 +185,18 @@ test("a run ending long after its wait is offered once, to a ready receiver, onl
     words: "Two agents are waiting.",
     origin: BRAIN_REQUEST_ORIGIN.TYPED,
   });
-  assert.equal(replies(c.thread(), runId)[0]?.words, "Two agents are waiting.");
+  assert.equal(c.replies(runId)[0]?.words, "Two agents are waiting.");
   // A duplicate claim, a repeated readiness, and a later report add nothing.
   assert.deepEqual(c.claim(offer), { granted: false });
   c.receiver.markReady(next);
-  await settle();
+  await drainMicrotasks();
   assert.equal(c.offers.length, 1);
-  assert.equal(replies(c.thread(), runId).length, 1);
+  assert.equal(c.replies(runId).length, 1);
 });
 
 test("a refused History write keeps the reply unoffered and ungranted until the write recovers, then it is offered once", async () => {
-  const c = composed();
-  const client = heldClient();
+  const c = brainHarness();
+  const client = heldModel();
   await c.host.replace(() => c.build(client));
   const agent = c.host.current();
   assert.ok(agent);
@@ -456,10 +209,10 @@ test("a refused History write keeps the reply unoffered and ungranted until the 
   });
   assert.equal(spoken.outcome, "accepted");
   if (spoken.outcome !== "accepted") return;
-  await settle();
+  await drainMicrotasks();
   c.refuse(true);
   client.release(answered("Done."));
-  await settle();
+  await drainMicrotasks();
   // Ended, but the thread refused the line: unmarked, not offered, and the
   // asking call is not granted the words either — nothing is said before
   // History holds it.
@@ -472,11 +225,11 @@ test("a refused History write keeps the reply unoffered and ungranted until the 
   // until the records next change — another ask accepted — and that report
   // writes the line, marks the run, and only then offers it, once.
   c.refuse(false);
-  await settle();
+  await drainMicrotasks();
   assert.equal(c.offers.length, 0);
-  await submitMany(c.submit, 1);
-  await settle();
-  assert.equal(replies(c.thread(), spoken.runId).length, 1);
+  await c.submitMany(1);
+  await drainMicrotasks();
+  assert.equal(c.replies(spoken.runId).length, 1);
   assert.deepEqual(
     c.offers.map((offer) => offer.runId),
     [spoken.runId],
@@ -492,18 +245,18 @@ test("a refused History write keeps the reply unoffered and ungranted until the 
 });
 
 test("two completions flushed together are offered one at a time, the second only after the first is acknowledged", async () => {
-  const c = composed();
-  const client = heldClient();
+  const c = brainHarness();
+  const client = heldModel();
   // Follow-up mode: two asks are two runs with two completions, where steer would answer both in one.
   await c.host.replace(() => c.build(client, { queueMode: QUEUE_MODE.FOLLOWUP }));
   const agent = c.host.current();
   assert.ok(agent);
-  const [runA, runB] = await submitMany(c.submit, 2);
+  const [runA, runB] = await c.submitMany(2);
   assert.ok(runA && runB);
   client.release(answered("Answer."));
-  await settle();
+  await drainMicrotasks();
   client.release(answered("Answer."));
-  await settle();
+  await drainMicrotasks();
   // Both ended with no receiver; readiness offers the oldest alone.
   const epoch = c.receiver.begin();
   c.receiver.markReady(epoch);
@@ -527,19 +280,19 @@ test("two completions flushed together are offered one at a time, the second onl
 });
 
 test("an offer the renderer never claimed is offered again to the next epoch; a claimed one is not, and old-epoch claims are refused", async () => {
-  const c = composed();
-  const client = heldClient();
+  const c = brainHarness();
+  const client = heldModel();
   await c.host.replace(() => c.build(client, { queueMode: QUEUE_MODE.FOLLOWUP }));
   const agent = c.host.current();
   assert.ok(agent);
   const first = c.receiver.begin();
   c.receiver.markReady(first);
-  const [runA, runB] = await submitMany(c.submit, 2);
+  const [runA, runB] = await c.submitMany(2);
   assert.ok(runA && runB);
   client.release(answered("Answer."));
-  await settle();
+  await drainMicrotasks();
   client.release(answered("Answer."));
-  await settle();
+  await drainMicrotasks();
   // A is offered and claimed; its renderer dies before the reply ends.
   const offerA = c.offers[0];
   assert.ok(offerA && offerA.runId === runA);
@@ -560,23 +313,23 @@ test("an offer the renderer never claimed is offered again to the next epoch; a 
   // refused even though the delivery ids match; the current epoch's is granted.
   assert.deepEqual(c.claim(offerB), { granted: false });
   assert.equal(c.claim(again).granted, true);
-  assert.equal(replies(c.thread(), runA).length, 1);
-  assert.equal(replies(c.thread(), runB).length, 1);
+  assert.equal(c.replies(runA).length, 1);
+  assert.equal(c.replies(runB).length, 1);
 });
 
 test("a Clear invalidates every delivery: an unclaimed offer is refused, a late acknowledgement is ignored, nothing is offered again", async () => {
-  const c = composed();
-  const client = heldClient();
+  const c = brainHarness();
+  const client = heldModel();
   await c.host.replace(() => c.build(client));
   const agent = c.host.current();
   assert.ok(agent);
   const epoch = c.receiver.begin();
   c.receiver.markReady(epoch);
-  const [runId] = await submitMany(c.submit, 1);
+  const [runId] = await c.submitMany(1);
   assert.ok(runId);
-  await settle();
+  await drainMicrotasks();
   client.release(answered("Before the clear."));
-  await settle();
+  await drainMicrotasks();
   assert.equal(c.offers.length, 1);
   const offer = c.offers[0];
   assert.ok(offer);
@@ -585,7 +338,7 @@ test("a Clear invalidates every delivery: an unclaimed offer is refused, a late 
   assert.deepEqual(c.claim(offer), { granted: false });
   assert.deepEqual(c.unclaimed(), []);
   await cleared;
-  await settle();
+  await drainMicrotasks();
   assert.equal(c.acknowledge(offer), false);
   // A new renderer reporting ready is offered nothing of the erased generation.
   c.receiver.reset();
@@ -594,21 +347,21 @@ test("a Clear invalidates every delivery: an unclaimed offer is refused, a late 
 });
 
 test("a launch that finds ended runs restores their words and speaks none of them", async () => {
-  const c = composed();
-  const client = heldClient();
+  const c = brainHarness();
+  const client = heldModel();
   await c.host.replace(() => c.build(client));
   const agent = c.host.current();
   assert.ok(agent);
-  await submitMany(c.submit, 2);
+  await c.submitMany(2);
   await c.host.replace(() => undefined);
-  await settle();
+  await drainMicrotasks();
   assert.equal(c.thread().filter((e) => e.kind === CONVERSATION_ENTRY_KIND.REPLY).length, 2);
   // The next launch: a fresh delivery owner, a fresh receiver, the same file.
-  const next = composed();
+  const next = brainHarness();
   next.storage.file = c.storage.file;
   next.receiver.markReady(next.receiver.begin());
-  await next.host.replace(() => next.build(heldClient()));
-  await settle();
+  await next.host.replace(() => next.build(heldModel()));
+  await drainMicrotasks();
   assert.equal(next.host.current()?.requests().length, 2);
   assert.equal(next.offers.length, 0);
   assert.deepEqual(next.thread(), []);
@@ -622,8 +375,8 @@ test("a launch that finds ended runs restores their words and speaks none of the
 test("a spoken ask's end is authorized exactly once across the call that asked and the offer path, whichever comes first", async () => {
   // Publication before the wait: the run ended, was written, marked, and
   // offered before the asking call's wait ever reached main.
-  const c = composed();
-  const client = heldClient();
+  const c = brainHarness();
+  const client = heldModel();
   await c.host.replace(() => c.build(client));
   const agent = c.host.current();
   assert.ok(agent);
@@ -636,9 +389,9 @@ test("a spoken ask's end is authorized exactly once across the call that asked a
   });
   assert.equal(spoken.outcome, "accepted");
   if (spoken.outcome !== "accepted") return;
-  await settle();
+  await drainMicrotasks();
   client.release(answered("Both are waiting."));
-  await settle();
+  await drainMicrotasks();
   assert.equal(c.offers.length, 1);
   const offer = c.offers[0];
   assert.ok(offer);
@@ -648,7 +401,7 @@ test("a spoken ask's end is authorized exactly once across the call that asked a
   assert.equal(waited.speak, true);
   assert.deepEqual(c.claim(offer), { granted: false });
   assert.equal(c.unclaimed().length, 0);
-  assert.equal(replies(c.thread(), spoken.runId).length, 1);
+  assert.equal(c.replies(spoken.runId).length, 1);
 
   // The reverse order: the offer is claimed first, and the wait is refused.
   const second = await c.submit({
@@ -658,9 +411,9 @@ test("a spoken ask's end is authorized exactly once across the call that asked a
   });
   assert.equal(second.outcome, "accepted");
   if (second.outcome !== "accepted") return;
-  await settle();
+  await drainMicrotasks();
   client.release(answered("Still waiting."));
-  await settle();
+  await drainMicrotasks();
   const secondOffer = c.offers[1];
   assert.ok(secondOffer && secondOffer.runId === second.runId);
   assert.equal(c.claim(secondOffer).granted, true);
@@ -678,9 +431,9 @@ test("a spoken ask's end is authorized exactly once across the call that asked a
   assert.equal(third.outcome, "accepted");
   if (third.outcome !== "accepted") return;
   c.acknowledge(secondOffer);
-  await settle();
+  await drainMicrotasks();
   client.release(answered("Later."));
-  await settle();
+  await drainMicrotasks();
   const staleWait = await c.waitOnCall(third.runId, epoch - 1);
   assert.equal(staleWait.speak, false);
   const thirdOffer = c.offers[2];
@@ -691,8 +444,8 @@ test("a spoken ask's end is authorized exactly once across the call that asked a
 });
 
 test("an on-call grant that takes the offered run out of the receiver's hand offers the next owed reply at once", async () => {
-  const c = composed();
-  const client = heldClient();
+  const c = brainHarness();
+  const client = heldModel();
   await c.host.replace(() => c.build(client));
   const agent = c.host.current();
   assert.ok(agent);
@@ -708,14 +461,14 @@ test("an on-call grant that takes the offered run out of the receiver's hand off
   });
   assert.equal(spoken.outcome, "accepted");
   if (spoken.outcome !== "accepted") return;
-  await settle();
+  await drainMicrotasks();
   client.release(answered("A."));
-  await settle();
-  const [typed] = await submitMany(c.submit, 1);
+  await drainMicrotasks();
+  const [typed] = await c.submitMany(1);
   assert.ok(typed);
-  await settle();
+  await drainMicrotasks();
   client.release(answered("B."));
-  await settle();
+  await drainMicrotasks();
   assert.deepEqual(
     c.offers.map((offer) => offer.runId),
     [spoken.runId],
