+++ apps/desktop/src/main/native/media-duck.test.ts
@@ -1,5 +1,6 @@
 import assert from "node:assert/strict";
 import test from "node:test";
+import { fakeNativeHelper } from "#testing/native-helper";
 import { MediaDuckController } from "./media-duck";
 import type { NativeHelperProcess } from "./native-helper";
 
@@ -9,46 +10,28 @@ const RELEASE_DELAY_MS = 40;
 interface Harness {
   controller: MediaDuckController;
   /** Every line written to every helper, in order, newlines stripped. */
-  commands: string[];
+  commands: () => readonly string[];
   spawns: () => number;
   stdinEnded: () => boolean;
   die: () => void;
 }
 
 function harness(options: { spawns?: (() => NativeHelperProcess | undefined)[] } = {}): Harness {
-  const commands: string[] = [];
   let spawned = 0;
-  let ended = false;
-  const exits: (() => void)[] = [];
+  const helper = fakeNativeHelper();
 
   const spawnHelper = (): NativeHelperProcess | undefined => {
     const scripted = options.spawns?.[spawned];
     spawned += 1;
-    if (scripted) return scripted();
-    return {
-      stdin: {
-        write: (chunk: string) => commands.push(chunk.trim()),
-        end: () => {
-          ended = true;
-        },
-      },
-      on: (event, listener) => {
-        if (event === "exit") exits.push(listener);
-      },
-      removeAllListeners: () => {
-        exits.length = 0;
-      },
-    };
+    return scripted ? scripted() : helper.process;
   };
 
   return {
     controller: new MediaDuckController({ spawnHelper, releaseDelayMs: RELEASE_DELAY_MS }),
-    commands,
+    commands: () => helper.written.map((chunk) => chunk.trim()),
     spawns: () => spawned,
-    stdinEnded: () => ended,
-    die: () => {
-      for (const exit of [...exits]) exit();
-    },
+    stdinEnded: helper.inputEnded,
+    die: helper.exit,
   };
 }
 
@@ -61,15 +44,15 @@ test("an exchange ducks at once and restores only after the hangover", async () 
   context.controller.setEnabled(true);
 
   context.controller.setExchangeActive(true);
-  assert.deepEqual(context.commands, ["duck"]);
+  assert.deepEqual(context.commands(), ["duck"]);
 
   context.controller.setExchangeActive(false);
   // The reply just ended; the players stay down through the pause a follow-up
   // question would arrive in.
-  assert.deepEqual(context.commands, ["duck"]);
+  assert.deepEqual(context.commands(), ["duck"]);
 
   await settle();
-  assert.deepEqual(context.commands, ["duck", "restore"]);
+  assert.deepEqual(context.commands(), ["duck", "restore"]);
 });
 
 test("a turn beginning inside the hangover keeps the duck held", async () => {
@@ -84,11 +67,11 @@ test("a turn beginning inside the hangover keeps the duck held", async () => {
   // Neither a restore — the conversation never stopped — nor a second duck,
   // SAFETY: Fixture value matches the narrowed runtime shape this test exercises.
   // which would read the ducked volume back as the one to restore to.
-  assert.deepEqual(context.commands, ["duck"]);
+  assert.deepEqual(context.commands(), ["duck"]);
 
   context.controller.setExchangeActive(false);
   await settle();
-  assert.deepEqual(context.commands, ["duck", "restore"]);
+  assert.deepEqual(context.commands(), ["duck", "restore"]);
 });
 
 test("disabled, an exchange never touches the helper at all", async () => {
@@ -100,7 +83,7 @@ test("disabled, an exchange never touches the helper at all", async () => {
   await settle();
 
   assert.equal(context.spawns(), 0);
-  assert.deepEqual(context.commands, []);
+  assert.deepEqual(context.commands(), []);
 });
 
 test("turning the setting off mid-duck restores at once, without the hangover", () => {
@@ -112,16 +95,16 @@ test("turning the setting off mid-duck restores at once, without the hangover", 
 
   // The hangover exists so a conversation does not pump; this is not a pause
   // in a conversation but the user asking for their volume back.
-  assert.deepEqual(context.commands, ["duck", "restore"]);
+  assert.deepEqual(context.commands(), ["duck", "restore"]);
 });
 
 test("enabling mid-exchange ducks the exchange already underway", () => {
   const context = harness();
   context.controller.setExchangeActive(true);
-  assert.deepEqual(context.commands, []);
+  assert.deepEqual(context.commands(), []);
 
   context.controller.setEnabled(true);
-  assert.deepEqual(context.commands, ["duck"]);
+  assert.deepEqual(context.commands(), ["duck"]);
 });
 
 test("a helper that dies is replaced for the next exchange", async () => {
@@ -134,11 +117,11 @@ test("a helper that dies is replaced for the next exchange", async () => {
   // restore, and the next exchange starts fresh.
   context.controller.setExchangeActive(false);
   await settle();
-  assert.deepEqual(context.commands, ["duck"]);
+  assert.deepEqual(context.commands(), ["duck"]);
 
   context.controller.setExchangeActive(true);
   assert.equal(context.spawns(), 2);
-  assert.deepEqual(context.commands, ["duck", "duck"]);
+  assert.deepEqual(context.commands(), ["duck", "duck"]);
 });
 
 test("enabling spawns the listening helper before any exchange", () => {
@@ -148,7 +131,7 @@ test("enabling spawns the listening helper before any exchange", () => {
   // The helper is up so it can hear the players' play-state broadcasts, but
   // nothing has asked for quiet, so nothing is written to it.
   assert.equal(context.spawns(), 1);
-  assert.deepEqual(context.commands, []);
+  assert.deepEqual(context.commands(), []);
 });
 
 test("a helper that cannot be spawned leaves the exchange unharmed", () => {
@@ -163,7 +146,7 @@ test("a helper that cannot be spawned leaves the exchange unharmed", () => {
   context.controller.setExchangeActive(false);
 
   assert.equal(context.spawns(), 2);
-  assert.deepEqual(context.commands, []);
+  assert.deepEqual(context.commands(), []);
 });
 
 test("a duck write that throws drops the helper, and the next exchange starts fresh", () => {
@@ -191,7 +174,7 @@ test("a duck write that throws drops the helper, and the next exchange starts fr
   context.controller.setExchangeActive(false);
   context.controller.setExchangeActive(true);
   assert.equal(context.spawns(), 2);
-  assert.deepEqual(context.commands, ["duck"]);
+  assert.deepEqual(context.commands(), ["duck"]);
 });
 
 test("a restore write that throws costs nothing but the dead helper", async () => {
@@ -223,7 +206,7 @@ test("a restore write that throws costs nothing but the dead helper", async () =
   // rather than the broken pipe.
   context.controller.setExchangeActive(true);
   assert.equal(context.spawns(), 2);
-  assert.deepEqual(context.commands, ["duck"]);
+  assert.deepEqual(context.commands(), ["duck"]);
 });
 
 test("stopping a controller whose pipe already broke does not throw", () => {
@@ -258,5 +241,5 @@ test("stopping closes stdin rather than writing, and writes nothing after", asyn
   // players back up on EOF — so a written restore would say it twice.
   assert.equal(context.stdinEnded(), true);
   await settle();
-  assert.deepEqual(context.commands, ["duck"]);
+  assert.deepEqual(context.commands(), ["duck"]);
 });
