+++ apps/desktop/src/testing/CLAUDE.md
@@ -1,0 +1,1 @@
+AGENTS.md
